shill: cellular: Use apn_type from modb on Cellular.APN
If the user selects an APN from the APN combobox, the APN is interpreted
as DEFAULT only, so the attach APN round robin does not prioritize this
APN in the Attach round robin. Fix this by honoring the apn_types value
from the modb when the value is prepolulated with modb values.

BUG=b:298473778
TEST=Verify that the correct Attach APN is used when cycling through
multiple attach APNs using the UI.

Change-Id: I6895fc69d54db78b23b16a74cff4c2e760d3b34c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4834093
Reviewed-by: Madhav <madhavadas@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>