libhwsec: Add HintUnknown to StackableError
This allows us to explicitly signal to the compiler that the state of an
error object is unknown. This is useful in some edge cases where we'd
otherwise have to disable the consumable warning or use much more
invasive workarounds like wrapping an error in optional<>.

This lets us remove a pragma where we disable the warning in cryptohome.
The reason the warning doesn't work without this as that the compiler
sees a loop that changes a CryptoStatus from "ok" to "unknown" which it
considers to be an error, but in this case we don't actually care about
the pre-loop state.

In theory I think it would actually be cleaner to be able to construct a
status in a deliberately unknown state instead, but in practice
declaring that is kind of ugly. The HintUnknown already aligns
reasonably well with the other Hint functions and is easy to use and
understand.

BUG=none
TEST=emerge-$BOARD cryptohome

Change-Id: Iae9e412b076343ad89b66d14c2e38bdf36e590ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4833656
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>