power: Power.BatterySaver.UserBrightenedSec metric
Introduce a metrics that logs the time it took the user to increase the
brightness above what battery saver dimmed it to.

So that we can tune the backlight dimming amount to a level that doesn't
cause a significant number of users to immediately brighten it.

BUG=b:298571460
TEST=Turn on battery saver, increase brightness, see histogram logged to
chrome://histogram, do it again, see it logged again.

Change-Id: Ie3fa2a21aabe2371f9a8ba7084d5d29972dc6fac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4835238
Tested-by: Charles Dick <cwd@google.com>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Charles Dick <cwd@google.com>