shill: Make link_name optional in LocalDevice
Currently, link_name is a mandatory constructor parameter for a
LocalDevice object. This works for HotspotDevice objects because shill
chooses a link_name when they are constructed, but it will not work for
upcoming P2PDevice objects, since P2PDevice object link_names are chosen
asynchronously by wpa_supplicant. Thus, we need to be able to construct
a LocalDevice without having a link_name.

To do this, make the link_name parameter an optional string. This
means that all consumers of LocalDevice objects need to account for the
possibility of link_name having no value. This is handled differently in
different places:

* LocalDevice:In places where link_name is logged, log
  "(no link name)" instead if we don't have a link name.

* HotspotDevice: The HotspotDevice constructor still uses a mandatory
  string argument for the link_name value, so a Hotspot device will
  always have a value inside link_name. This class just directly
  accesses link_name.value(), since it will always be populated (and if
  it isn't populated for some reason, we're in a totally unexpected
  state that we don't know how to recover from).

* TetheringManager: Expects hotspot_dev_->link_name to have a value at
  all times, and uses CHECKs to assert this.

* WiFiProvider: Uses CHECKs to assert that a LocalDevice has a link name
  upon registration and deregistration.

BUG=b:295060345
TEST=FEATURES=test emerge-guybrush shill

Change-Id: I5c4b3ba4a42de0e97ea94a4e84c66dab2368a468
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4834144
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>