vm_tools: trim user filesystem on vm startup on lvm devices
If discards on the host can sometimes get interrupted and lost.
Causing spaced to sometimes have an incorrect view of the world.
Because storage ballooning uses spaced to operate, this can
drastically restrict functionality of vms which use storage
ballooning. We now run fstrim on vm startup (for vms which
use ballooning) to reduce this problem. For scenarios where
no work is to be done, it should essentially be a no-op.

BUG=b:296136986
TEST=on nissa (lvm) and volteer (non-lvm)

Change-Id: Id17e7e572a95135900c54d517d9d1dbece101fa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4837963
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: Daniel Ng <danielng@google.com>
Commit-Queue: Daniel Ng <danielng@google.com>