sepolicy: Permit cryptohome to remove socket by doing a getattr
When crostini is setup, it creates a socket with
cros_home_shadow_uid_root_crosvm label which is not accessible by
cryptohome. When the user is being removed, cryptohome tries to call
getattr on this, which creates an SELinux violation logged in audit.log.
The violation prevents the user cryptohome directory to be removed
completely, which in turn tells chrome that user still exists. This results in the user pod being shown, but the user cannot actually log in.

Note: This CL prevents future occurrences of this error and does not
do anything for prior occurrences.

BUG=b:283183030
TEST=Setup crostini, remove user and readd user.
TEST=test in chromium:4839221 should pass

Change-Id: I203b40860b3e0506a954913203ce2a78942f4ae5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4840374
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>