cryptohome: Create limiter during Prepare
As a part of the upcoming refactor, we'll move some of the
AddFingerprintAuthFactor steps to PrepareAuthFactor, to allow more
latency optimization. For this, we need to create rate-limiters not on
the first add, but on the first prepare-for-add. Incorporating the
parameters to create a rate-limiter into Prepare is complicated, and the
fact that we might move this operation around (like this time) in the
future again probably means that it should be an isolated API.

BUG=b:298346899
TEST=cryptohome unittests
TEST=tast run cryptohome.FingerprintManual

Change-Id: Ia57f0e48d9c7cb0bdfdf8c2d23ec81995c8f2a04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4840801
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>