cryptohome: Do less AuthFactor on the heap
The AuthFactor object is just a value object that can be copied and
moved so storing all of them in a unique_ptr so that they all have to go
in the heap is kind of unnecessary.

This change rewrites AuthFactorMap to just store the factors directly,
and accept them by value instead of by unique_ptr. It also tries to
track down all of the other AuthFactor on the heap and fix them up to
match.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Iddd7e587c3978466c651cfb4553b21bf3b99a956
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4844399
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>