power: Workaround battery sustainer issue
This works around a battery sustainer issue where the battery won't
discharge if it's full, even if the upper limit for the battery
sustainer is lower than the current charge. It does this by force
discharging the battery for 1 second if the battery is full when
enabling Charge Limit. This is sufficient to start the charging task on
the EC, and properly trigger the battery sustainer when it is enabled.

BUG=b:295240206
TEST=tast.power.AdaptiveCharging

Change-Id: I1fa7e1905bfc204cad37921068bfdc312ae3af39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4811788
Auto-Submit: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Yanye Li <yanyeli@google.com>