vm_tools: Naively async-ify the concierge service's init
We make service's init "async" in the most naive way possible. This is
still a blocking synchronous call, but it looks async from the
perspective of the caller (which will still be blocked by it).

We do this separately to actually making initialization async in order
to separate the api-changes from the implementation details.

BUG=b:296025701
TEST=tast tests

Change-Id: I642c9b5faf0f873da1fe3d7619f52c1cc65958ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4845448
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>