multiple: Refactor cryptohome FP flow
Switch biod_proxy and cryptohome to use the updated FP API interface.
The major change is that GSC communication for rate-limitation is now
moved to the Prepare step to reduce latency of the actual auth.

The retry logic of fingerprint auth session is removed in this CL
because it'll need to be reworked. It will be added as a follow-up.

BUG=b:298346899
TEST=pending
TEST=fingerprint login integration tests fail

Change-Id: If7194635bc8d31d3479e8ed486ec6d94fecf7af1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4845668
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>