shill: wifi: Add interface combination checking
To fully support concurrency, it is necessary to have not only static
concurrency reporting capabilities, but also a dynamic system, that
would determine what interface type *yet* can be brought up and in
which mode (MCC/SCC) can it happen.

It is based on what are possible interface combinations reported
by the driver and which interface types are in use at the moment.

Also in order to avoid race conditions between different asynchronous
requests, the determination of available interface types should
be performed in "reserve/free" mode. In this mode, caller before
starting to bring up the interface should ask WiFiPhy to reserve the use
of certain interface type, then free it when finished using that
interface (or failed to set it up!).

The code is written in a forward-compatible manner, to minimize changes
necessary for using it in the likely future scenarios, e.g. AP+P2P
or P2P Discovery.

BUG=b:299295629
TEST=shill UT

Change-Id: I9b17aab6e8df7a1e8c7cef2d5654a2c8af18400f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4846557
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Jacek Siuda <jsiuda@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>