crash: Collect device model if USE=hw_details
Crashes typically include the model details in the form of the hwclass.
For ChromeOS Flex there are hundreds of models with the same hwclass,
which can make debugging difficult.

If the hw_details USE flag is set, include the product name,
product version, and system vendor dmi fields in the crash metadata.

We store these in keys that match those in feedback reports to make it
clearer that they contain the same data.

BUG=b:299149315, b:274669186
TEST=`cros_workon_make -breven --test crash-reporter`
TEST=run in a vm, find a .meta file in /var/spool/crash, see new fields

Cq-Depend: chromium:4845698
Change-Id: I7cb8eb1d6fcced0c06634a9dffa81b8960cf4a76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4847491
Commit-Queue: Ted Brandston <tbrandston@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>