common-mk: disable -Wimplicit-int-float-conversion
As detailed in b/296451257#comment2, this warning has been a nop since
day 1. Other projects which we try to be (at least somewhat) consistent
with - like Chromium and Google3 - either ignore this flag, or outright
disable it.

If individual projects want to live with this, that's great; forcing
platform2 to live with it by default seems a bit questionable, though.

BUG=b:296451257
TEST=build_packages --board=brya with a modified compiler wrapper

Change-Id: I0c6ae998d13b0f896c97b1b527c1af02117da971
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4846422
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>