cryptohome: Add DecryptedUss class
This initial version of the class is used a bit strangely since this is
trying to incrementally replace UserSecretStash rather than do it all in
one huge change.

For now the only logic moved out of UserSecretStash is the decryption
logic. Thus, the class now basically loads a DecryptedUss, rips the
contents out of it, and then throws it away. Once some support for doing
write operations is added this should be unnecessary but doing things
this way for now allows the change to be a little more incremental.

BUG=b:295898583
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I3448f6865a26543065696c96082af1a69a117bbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4847364
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>