power: Fix flag set when Charge Limit is enabled
is_sustainer_set is incorrectly set to false if the sustainer is left on
when Charge Limit is enabled when StopAdaptiveCharging is called.
StopAdaptiveCharging is still called, even when it isn't enabled, to
report metrics on how well the ML model did in predicting when the user
will unplug.

BUG=b:280880357
TEST=unittests

Change-Id: I9cb5bef2f83c73a74e90977c0cf1466f0d0c7d6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4881027
Reviewed-by: Yanye Li <yanyeli@google.com>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Auto-Submit: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>