cryptohome: Add Reprepare PostAction
From Chrome perspective, the flow of dealing with auth factors that need
prepare is to:
1. PrepareAuthFactor
2. AuthenticateAuthFactor based on signal, may retry and happen several
   times.
3. TerminateAuthFactor to make the factor not available anymore.

However, as an implementation detail, the asynchronous steps of a
fingerprint auth factor authentication needs to happen "per auth",
unlike the legacy fingerprint flow. We'll address this with a
"Reprepare" PostAction which will prepare the auth factor again after an
authentication.

BUG=b:298346899
TEST=pending
TEST=cryptohome.FingerprintManual passes, and latency improves largely

Change-Id: Iaee0c65c26c58309b1407cc7199454d6df030372
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4852668
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>