patchpanel: Refactor MinijailedProcessRunner::RunSyncDestroy()
In the current code, when RunSyncDestroy() needs the output of stdout
or stderr from the child process, it will wait for process to exit at
first, and then consume the whole buffer in the pipes.

This patch reimplements this logic using poll(): the function will
poll() the pipes of stdout and stderr, until both pipes return the
POLLHUP events, which indicates the process is exiting. After that, the
function call waitpid() to wait for process to exit and clean up
resources.

The advantage of the new code are:
- Read the pipes when the child process is running. This can avoid the
  case that the pipe buffer is full and the child process can be blocked
  by writing to this buffer.
- Make it easier to add timeout logic to limit the run time of the child
  process.

This patch shouldn't have any observable change, execpt for the log
messages are also refactored a little bit.

BUG=b:175842822
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=Manually checked that traffic counters works well; net.log has the
  expected iptables error logs.

Change-Id: I72306769e8dfc6b501130269e0cfb2415a225cee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4852545
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>