lorgnette: Refactor command handling in cli
lorgnette_cli is getting a longer and longer list of supported commands.
Instead of adding string comparisons in multiple places, switch to an
exhaustive enum and map lookup.  The set of accepted commands and
behavior doesn't change in this CL, but it will be easier to add future
commands.

BUG=b:275043885
TEST=Run lorgnette_cli on a device

Change-Id: I9f80ddcfd5c15f1cdb185cbbb23feee28aa14219
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4854065
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>