vm_tools: Fix misreporting of shader cache install success
Since DLC installation is not guaranteed to be synchronous, the current
logic is only correct when waiting.

Add a success message to processing of the ShaderCacheMountStatusChanged
message and ensure a callback is always registered. This will serve as a
notification of success for non-waiting calls that would otherwise miss
this log entry as a result of the change in garcon.

Lastly, ensure that the installation of shader cache DLCs are at least
attempted even if cicerone cannot wait for a success notification. This
should improve user experience in this corner case since DLC
installation may still succeed and provide cache hits.

BUG=b:299526066
TEST=Deploy to device and observe logs.

Change-Id: I39e48dbc11c19f64bd1b7ca27ed72c96e6227aef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4852443
Commit-Queue: Renato Pereyra <renatopereyra@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Tested-by: Renato Pereyra <renatopereyra@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>