lorgnette: Add cli advanced_scan verb
The lorgnette_cli advanced_scan verb uses the new StartPreparedScan and
ReadScanData APIs to fetch scan data.  If the source appears to be a
document feeder, it continues fetching pages until the feeder is empty;
otherwise, it scans a single page.  The naming convention for output
files from the existing scan verb is reused for advanced_scan.

SetOptions is not yet implemented in lorgnette, so any scans use
the scanner's default options.

BUG=b:275043885
TEST=lorgnette_cli advanced_scan on various scanners

Change-Id: I4f65a6e02ad6d0af4512e11d60ea3f47c023b6ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4876316
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>