arc: keymint: Add Root of Trust
This CL adds the Root of Trust capability to KeyMint.
It does so by adding the appropriate conversion
functions for root of trust parameters.

Also, it adds a new function
method in KeyMint Server specifically for Get
Root Of Trust method. This is required as the template
function could not be used as required response
could not fit into it.

BUG=b:289270498
TEST=emerge-brya chromeos-base/arc-keymint
TEST=cros_workon_make --board=brya chromeos-base/arc-keymint --test

Change-Id: Ib501e35345657e551a1e58fa41cb6f34488551d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4855208
Tested-by: Vaibhav Raheja <vraheja@chromium.org>
Reviewed-by: Yao Li <yaohuali@google.com>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>