ml: Add the heatmap model v0 to ml service.
This model is to be used on Rex device with real heatmap input.
We will remove the old test model after we update the chromium code.
The new model size is 90KB while the old one is 30KB, mostly because the rex device has a bigger screen than tangor.

BUG=b:289735201
TEST=unit test

Change-Id: I279ff9cd159573b6bbad607fca89d6ec6aa4693c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4861324
Commit-Queue: Jing Wang <jiwan@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Tested-by: Jing Wang <jiwan@chromium.org>