cryptohome: Keep AuthSession in use until op completes
Currently some of the operations don't bind the InUseAuthSession to the
callback passed to the AuthSession. This means the InUseAuthSession will
be destructed as soon as the entry function returns, while the actual
operation is still running. We should ensure the InUseAuthSession is
always bound as callback argument.

BUG=b:299872991
TEST=cryptohome unittests

Change-Id: I37021f9d05cf1fe20c722908489343c51dd4784e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4852796
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>