cryptohome: Add Run(AuthSession)WhenAvailable
Currently if the auth session requested is in use, an error will be
returned. This is not convenient for the caller in some cases. As dbus
requests itself is already in async manner, an improvement we can make
is to queue upcoming operations on an active auth session, and process
them sequentially.

This CL doesn't change the said behavior yet, it only adds an API in
auth_session_manager that supports the said behavior.

BUG=b:299872991
TEST=cryptohome unittests

Change-Id: I0f97432cbf057008fb3ddb15be32e76bab808fa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4856578
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>