cryptohome: Add DecryptedUss::Transaction, use it internally
This basically shifts all of the UserSecretStash mutators into a new
Transaction inner class for DecryptedUss, and then under the covers of
UserSecretStash it basically replaces all of the internal guts with
calls to DecryptedUss code.

The next step would then be to get rid of UserSecretStash entirely and
replace all the code calling it with direct usage of DecryptedUss (and
Transaction) instead, but this changes doesn't include that step since
it's already rather large.

One significant change it does make to all the client code right away
though is to eliminate the main key as an external parameter. The
existing code has the key get returned by the loading process and then
it gets passed back into USS for the operations that require it. This
change just keeps the main key internally in DecryptedUss and so the
calling code never sees it directly. The idea was mostly to simplify the
commit process but overall I think it's actually preferable to minimize
how much main key gets passed around. It also makes it impossible to
have certain types of bugs (e.g. adding two different wrapped main
keys).

Note that there is a bit of a performance hit from this change: on every
mutating operation we serialize & re-encrypt the USS in memory. So if
you have a change which does something like "add a wrapped key and reset
secret" it now does two encrypts. However, given that mutating operation
are relatively rare this shouldn't be a serious problem and it's only
transitional as once the next change replaces UserSecretStash entirely
with DecryptedUss that cost will go away again.

BUG=b:295898583
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ieb1496290b07c6efe6d37804d6a7c430274752a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4868555
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>