missive: Fix debug on/off flag attachment.
We will now always consume health_data_logging flag (true or false)
if it is present in the proto (coming from Ash), and not change
anything otherwise (coming from LaCros or other daemons).

Chrome is going to be changed accordingly too.

While investigating why some debug records are missing,
I came to realize that this was what prior code intended:
it treated missing flag as false. Therefore if a proto
was coming not from Ash, it always turned off history.

In addition, Recorder for upload is destructed when the upload
payload is ready - that way its outcome will be included in
the health history update being sent with the upload already.

BUG=b:214041435
TEST=cros_sdk cros_workon_make --board=${BOARD} missive --test
TEST=manual VM testing

Change-Id: I9032a8f1ac8d056cada041c8227d6c53da47b359
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4857698
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Hong Xu <xuhong@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>