crash: Add missing `inline` to `constexpr char[]`
Minor fixup to `kLacrosVariationsListFile`. All the other constant
c-strings in paths.h are declared inline, which ensures that only a
single copy of the string is used.

It wasn't immediately clear why this one was being treated differently,
thanks to eshimanovich@ for pointing out that when that was added in https://crrev.com/c/4632977/1 `inline` hadn't been added to that file
(that came in https://crrev.com/c/4656097, which was developed in
parallel).

BUG=None
TEST=CQ

Change-Id: I8766583c71080bf55fcd81728c4d3923b10ad788
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4858859
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>