flex_hwis: Simplify test data creation
The current implementation repeatedly returns a TelemetryInfoPtr that's
also held as a member. It's a slightly confusing interface that results
in odd functions like MockTelemetryInfo which, in effect, repeatedly
std::move a member variable into itself, and makes it very easy to
accidentally lose/use-after-free the held TelemetryInfoPtr.

Rather than returning it from every function, we can have a `Get`
method to retrieve a copy of the TelemetryInfoPtr when it's needed.

Also rename the methods to fit the new interface better.

BUG=None
TEST=cros_workon_make -breven --test chromeos-base/flex_hwis

Change-Id: I816ac082798f61efd7119fca335a3bf84aa1bb72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4858907
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>
Auto-Submit: Ted Brandston <tbrandston@google.com>