init: startup: Outside ChromeosStartup: Use crossystem by reference
It is safer to pass reference to objects instead of their address.
Although the contract is the caller promises the reference is valid
during the function call, |mount_helper_factory| was keeping a reference
around and use it when |Generate()| is called.
|Generate()| is now called with a reference to |crossystem| and is
not keeping it.

BUG=b:233416937
TEST=Unit tests.

Change-Id: Id73706f8a348a71ac6f60c04abe60d2cb3478b94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4861436
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>