shill: cellular: use longer timeout when tethering requires a new PDN connection
If the tethering operation involves connecting a new PDN connection
(e.g. DUN as DEFAULT) request the tethering manager to use a longer
timeout for the attempt (45s) Otherwise (e.g. reusing DEFAULT), use a
much shorter timeout (10s).

The longer timeout is requested by the cellular device as soon as it
knows the type of procedure that needs to be run in the tethering
attempt. Therefore, all the attempts will start with the shorter
timeout and only be updated if explicitly requested by the cellular
device.

BUG=b:278026029
TEST=Manual testing

Change-Id: Ib7330bbc533f37211fad4a5faf5c9e30940d1d2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4859643
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>