hiberman: Report metric of pages with zeroes in the hibernate image
The kernel doesn't include pages that only contain zeroes in the
hibernation image. The hibernation image keeps track of these pages,
they are restored when the image is loaded for resuming.

Report a metric with the percentage of pages with zeroes in the
hibernate image in memory. The value is calculated like this:

  pages_with_zeroes / (no_zero_pages + pages_with_zeroes)

BUG=b:300463848
TEST=hibernate and resume
  open chrome://histograms/DroppedPagesWithZeroesPercent

Change-Id: I20201b5223e4faa4baa2a5a62a0faf9d39508e54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4867067
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>