shill: wifi: do not use cellular region hint which is ignored in Linux-wireless
When setting WiFi region, NL80211_ATTR_USER_REG_HINT_TYPE will be
default set to NL80211_USER_REG_HINT_USER if this attribute is not
present. Shill set this to NL80211_USER_REG_HINT_CELL_BASE when starting
SoftAP which is ignored in Linux-wireless __reg_process_hint_user.
Remove this attribute so that region could be processed in the kernel.

BUG=b:298197280
TEST=FEATURES=test emerge shill
TEST=on corsola(MT7921), start tethering, check the region is set. Check
that the hotspot could be enabled in 5GHz band.
TEST=on corsola(MT7921), stop tethering, check the region is reset back
to WORLD.
TEST=on corsola(MT7921), connect WiFi to an AP. Start tethering success
after that.

Change-Id: I3f291efc18f5fdeb132313dcc96576a92c5743ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4867101
Tested-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>