lorgnette: Move file ownership out of ImageReader
ImageReader currently takes ownership of the passed-in FILE* and closes
it when the encoded image is fully written.  This works fine for the
case where a d-bus client passes in an open fd, but doesn't allow
ImageReader users to do anything more advanced.  To support the new
ReadScanData managing its own FILE*, move ownership out of ImageReader
and make the existing uses rely on a ScopedFILE in the calling context.
Because the file is no longer closed at the end of writing the image,
call fflush to ensure all data is written.

BUG=b:297443322
TEST=CQ

Change-Id: Icb21e6dab05edbf63355e778a040540ffc7f7dda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4875035
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>