cryptohome: Handle invalid fd in the oop mount helper polling
When the WaitForHelper call in mount is given an invalid file descriptor
it will now fail immediately rather than waiting for the poll() to fail.
This is technically the same in terms of error behavior, but doing the
easy check up front saves waiting 12s on every call.

This change is more to make the unit test that triggers this code fast,
in the wild this specific failure mode isn't likely to happen.

BUG=b:300519536
TEST=the unit tests don't take 12+ seconds

Change-Id: I8f4aee29eb3580934d6b669dc9876a33e01def0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4867107
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>