lorgnette: Implement ReadScanData
ReadScanData is the pull-based equivalent to the previous push-based
GetNextImage.  It reads a chunk of data from the scanner, encodes it in
the client's requested encoding, and returns the next chunk of encoded
image data to the client.

Most of the actual reading and encoding is handled by SaneDevice, so
this CL primarily just hooks it up to the d-bus methods.

BUG=b:297443322
TEST=tast run scanner.AdvancedScan.*

Change-Id: I16a574efcac0719f316268addb5eb32343d04e92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4876315
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>