cryptohome: Add Commit with storage
This adds an overload to the DecryptedUss transaction Commit that also
includes writing the value out to persistent storage as part of the
commit. This will be needed to actually switch the existing AuthSession
code over to using DecryptedUss since it'll need a way to do the
in-memory and in-storage commits as a single operation.

Ideally we wouldn't need both Commit overloads, but at least until the
transition to the new classes are done it's hard to get by without the
memory-only version. After that transition we can hopefully drop the
redundancy.

BUG=b:295898583
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I2af5b76ffe609c7fc7b5a7ecad1eec5765dcfa0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4869141
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>