system_api: Add lorgnette ReadScanData progress field
The older lorgnette push API sends back a completion percentage with
each progress signal.  This was overlooked in the new design, so callers
currently have no way to display a progress bar.

Add a percentage completion field to the ReadScanDataResponse field so
d-bus clients can have this feature with the new API as well.  Progress
is returned as a field on ReadScanData rather than an up-front line
count because future formats may not have a meaningful number of lines
(e.g. PDF).

The field is an integer from 0-100 rather than a float even though
lorgnette internally can calculate a more precise percentage.  This is
done because scanners typically return more than 1% of a file in a
single sane_read call, so the count potentially jumps by multiple
percent on each ReadScanData call.  Adding more significant digits makes
it look more precise in the protobuf, but doesn't increase the client's
ability to show a meaningful progress interval.

BUG=b:297443322
TEST=cros build-packages

Change-Id: Ie1b07248462152e24e0157808971721596618665
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4876314
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>