crash: handle Linux 6.x BTRFS corruption message
The BTRFS tree corruption log message has changed slightly again with
Linux 6.0, similar to the previous b/262691354 from Linux 5.15. Update
the regex to match the new pattern and add another test case.

Specifically, the difference is in this part of the log message:

5.15: checksum verify failed on 122798080 wanted ...
6.0:  checksum verify failed on logical 122077184 mirror 1 wanted ...

<https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=8f0ed7d4e7bd87c9207a59d6d887777f632a5ed5>

This is needed since the Crostini (project-termina) guest kernel is
switching to Linux 6.1.

BUG=b:300849008
BUG=b:299353857
TEST=FEATURES=test emerge-brya crash-reporter

Change-Id: I0282cd24a4a1d7fb0cff1ca2d14ba1adeb506d47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4874837
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>