cryptohome: Add UserUssStorage
This adds a small wrapper around UssStorage that basically binds it to a
specific username. This is helpful in the context of AuthSession where
you always have only a specific user, allowing us to cut down a bunch of
cases where you have to pass around two parameters and the username is
only being supplied because it's a required parameter for all the
UssStorage functions.

BUG=b:295898583
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I923afc282119c52542d91ded3161dd83e263c338
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4874724
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>